separate endpoint for OTON
