+++ src/server/operandi_server/routers/__init__.py
@@ -1,6 +1,7 @@
 __all__ = [
     "RouterAdminPanel",
     "RouterDiscovery",
+    "RouterOton",
     "RouterUser",
     "RouterWorkflow",
     "RouterWorkspace"
@@ -8,6 +9,7 @@ __all__ = [
 
 from .admin_panel import RouterAdminPanel
 from .discovery import RouterDiscovery
+from .oton import RouterOton
 from .user import RouterUser
 from .workflow import RouterWorkflow
 from .workspace import RouterWorkspace
