--- src/server/operandi_server/routers/constants.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-
-class ServerApiTags(str, Enum):
-    ADMIN = "admin"
-    DISCOVERY = "discovery"
-    USER = "user"
-    WORKFLOW = "workflow"
-    WORKSPACE = "workspace"
