+++ src/server/operandi_server/server.py
@@ -11,7 +11,8 @@ from operandi_utils.database import db_initiate_database
 from operandi_utils import safe_init_logging
 
 from operandi_server.files_manager import LFMInstance
-from operandi_server.routers import RouterAdminPanel, RouterDiscovery, RouterUser, RouterWorkflow, RouterWorkspace
+from operandi_server.routers import (
+    RouterAdminPanel, RouterDiscovery, RouterOton, RouterUser, RouterWorkflow, RouterWorkspace)
 from operandi_server.server_utils import (
     insert_default_accounts, insert_production_workflows, produce_production_workflows)
 
@@ -103,6 +104,7 @@ class OperandiServer(FastAPI):
     async def include_webapi_routers(self):
         self.include_router(RouterAdminPanel().router)
         self.include_router(RouterDiscovery().router)
+        self.include_router(RouterOton().router)
         self.include_router(RouterUser().router)
         self.include_router(RouterWorkflow(self.production_workflows).router)
         self.include_router(RouterWorkspace().router)
